Refactor the service to calculate the patients age
