+++ spec/services/patients/age_spec.rb
@@ -1,11 +1,11 @@
 require 'timecop'
-require_relative '../../app/services/age_calculator'
+require_relative '../../../app/services/patients/age'
 
-describe AgeCalculator do
+describe Patients::Age do
   context 'when date is *NOT* passed' do
     it 'calculates age in years and months' do
       Timecop.freeze('2011-07-29') do
-        age = described_class.calculate(Date.new(1979, 9, 10))
+        age = described_class.from(Date.new(1979, 9, 10))
         expect(age.years).to eq(31)
         expect(age.months).to eq(11)
       end
@@ -14,7 +14,7 @@ describe AgeCalculator do
 
   context 'when date is passed' do
     it 'calculates age in years and months' do
-      age = described_class.calculate(Date.new(1979, 9, 10), Date.new(2011, 7, 29))
+      age = described_class.from(Date.new(1979, 9, 10), Date.new(2011, 7, 29))
       expect(age.years).to eq(31)
       expect(age.months).to eq(11)
     end
